refactor: route handler が features/interface を経由するよう依存境界を是正
- reports route が application を直 import していたのを interface 経由に統一
  （features/reports/interface/commands を新設、blog/views と同じ経路に揃える）
- cron route（sync-articles / sync-baseline）が push-notification の infrastructure を
  直叩きしていたのを interface 経由に修正
  （features/push-notification/interface/commands を新設）

注: admin の application 層からの @repo/database 直 import を infrastructure へ寄せる件は、
sync-articles のテストが db を直接モックしている現行慣習との整合・テスト大幅改修を伴うため、
oxc カスタムルール導入と合わせた別タスクとする。
